+++ examples/07 - jinja_templates/hello.py
@@ -7,7 +7,7 @@ from common import WINDOW_SIZE
 
 import glue
 
-glue.init('web')                     # Give folder containing web files
+glue.init()                     # Serve the ui/ folder
 
 @glue.expose
 def py_random():