+++ examples/09 - disable_cache/disable_cache.py
@@ -6,8 +6,8 @@ from common import WINDOW_SIZE
 
 import glue
 
-# Set web files folder and optionally specify which file types to check for glue.expose()
-glue.init('web')
+# Serve ui/ (default) and optionally specify which file types to check for glue.expose()
+glue.init()
 
 # disable_cache now defaults to True so this isn't strictly necessary. Set it to False to enable caching.
 glue.start('disable_cache.html', size=WINDOW_SIZE, disable_cache=True)